Actualizaciones YOY, tienda, about, contact, manager

+++ system/libraries/Session/drivers/Session_files_driver.php
@@ -383,7 +383,11 @@ class CI_Session_files_driver extends CI_Session_driver implements SessionHandle
 				continue;
 			}
 
-			unlink($this->_config['save_path'].DIRECTORY_SEPARATOR.$file);
+			//unlink($this->_config['save_path'].DIRECTORY_SEPARATOR.$file);
+                        if( is_file($this->_config['save_path'].DIRECTORY_SEPARATOR.$file) )
+                        {
+                            unlink($this->_config['save_path'].DIRECTORY_SEPARATOR.$file);
+                        }
 		}
 
 		closedir($directory);
